+++ Tests/Zoo.psm1
@@ -1,0 +1,8 @@
+﻿
+filter Out-Table {
+	$_.ToTable() | Out-Host
+}
+
+function Convert-NewLine([string]$text) {
+	$text.Replace("`r`n", "`n")
+}